webharvest/worker/requirements.txt
# Core Dependencies
celery[redis]==5.3.4
redis==4.6.0
flower==2.0.1  # Celery monitoring

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Web Scraping
playwright==1.40.0
beautifulsoup4==4.12.3
lxml==5.1.0
readability-lxml==0.8.1
markdownify==0.11.6
html2text==2020.1.16

# HTTP Clients
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# Content Processing
Pillow==10.2.0  # Image processing
PyPDF2==3.0.1  # PDF extraction
python-magic==0.4.27  # File type detection

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
urllib3>=1.26.14,<2.0.0

# Monitoring & Logging
structlog==24.1.0
prometheus-client==0.19.0

# Vector Database
qdrant-client==1.7.0
sentence-transformers==2.2.2  # For embeddings

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0